+++ lib/index.js
@@ -18,140 +18,15 @@ var __importStar = (this && this.__importStar) || function (mod) {
     __setModuleDefault(result, mod);
     return result;
 };
-Object.defineProperty(exports, "__esModule", { value: true });
-const React = __importStar(require("react"));
-const react_1 = require("react");
-const react_native_1 = require("react-native");
-const react_native_webview_1 = require("react-native-webview");
-const helper_1 = require("./helper");
-const CLOSE_URL = 'https://standard.paystack.co/close';
-const Paystack = ({modal = true, style={}, paystackKey, billingEmail, phone, lastName, firstName, amount = '0.00', currency = 'NGN', channels = ['card'], refNumber, billingName, subaccount, handleWebViewMessage, onCancel, autoStart = false, onSuccess, activityIndicatorColor = 'green', }, ref) => {
-    const [isLoading, setisLoading] = react_1.useState(true);
-    const [showModal, setshowModal] = react_1.useState(false);
-    const webView = react_1.useRef(null);
-    react_1.useEffect(() => {
-        autoStartCheck();
-    }, []);
-    react_1.useImperativeHandle(ref, () => ({
-        startTransaction() {
-            setshowModal(true);
-        },
-        endTransaction() {
-            setshowModal(false);
-        },
-    }));
-    const autoStartCheck = () => {
-        if (autoStart) {
-            setshowModal(true);
-        }
-    };
-    const refNumberString = refNumber ? `ref: '${refNumber}',` : ''; // should only send ref number if present, else if blank, paystack will auto-generate one
-    const subAccountString = subaccount ? `subaccount: '${subaccount}',` : ''; // should only send subaccount with the correct subaccoount_code if you want to enable split payment on transaction
-    const Paystackcontent = `   
-      <!DOCTYPE html>
-      <html lang="en">
-        <head>
-          <meta charset="UTF-8">
-          <meta http-equiv="X-UA-Compatible" content="ie=edge">
-          <meta name="viewport" content="width=device-width, initial-scale=1.0">
-          <title>Paystack</title>
-        </head>
-          <body  onload="payWithPaystack()" style="background-color:#fff;height:100vh">
-            <script src="https://js.paystack.co/v2/inline.js"></script>
-            <script type="text/javascript">
-              window.onload = payWithPaystack;
-              function payWithPaystack(){
-              var paystack = new PaystackPop();
-              paystack.newTransaction({ 
-                key: '${paystackKey}',
-                email: '${billingEmail}',
-                firstname: '${firstName}',
-                lastname: '${lastName}',
-                phone: '${phone}',
-                amount: ${helper_1.getAmountValueInKobo(amount)}, 
-                currency: '${currency}',
-                ${helper_1.getChannels(channels)}
-                ${refNumberString}
-                ${subAccountString}
-                metadata: {
-                custom_fields: [
-                        {
-                        display_name:  '${firstName + ' ' + lastName}',
-                        variable_name:  '${billingName}',
-                        value:''
-                        }
-                ]},
-                onSuccess: function(response){
-                      var resp = {event:'successful', transactionRef:response};
-                        window.ReactNativeWebView.postMessage(JSON.stringify(resp))
-                },
-                onCancel: function(){
-                    var resp = {event:'cancelled'};
-                    window.ReactNativeWebView.postMessage(JSON.stringify(resp))
-                }
-                });
-                }
-            </script> 
-          </body>
-      </html> 
-      `;
-    const messageReceived = (data) => {
-        const webResponse = JSON.parse(data);
-        if (handleWebViewMessage) {
-            handleWebViewMessage(data);
-        }
-        switch (webResponse.event) {
-            case 'cancelled':
-                setshowModal(false);
-                onCancel({ status: 'cancelled' });
-                break;
-            case 'successful':
-                setshowModal(false);
-                const reference = webResponse.transactionRef;
-                if (onSuccess) {
-                    onSuccess({
-                        status: 'success',
-                        transactionRef: reference,
-                        data: webResponse,
-                    });
-                }
-                break;
-            default:
-                if (handleWebViewMessage) {
-                    handleWebViewMessage(data);
-                }
-                break;
-        }
-    };
-    const onNavigationStateChange = (state) => {
-        const { url } = state;
-        if (url === CLOSE_URL) {
-            setshowModal(false);
-        }
-    };
-    return modal === true ?
-        (<react_native_1.Modal style={[ { flex: 1 }, style ]} visible={showModal} animationType="slide" transparent={false}>
-            <react_native_1.SafeAreaView style={{ flex: 1 }}>
-                <react_native_webview_1.WebView style={[ { flex: 1 } ]} source={{ html: Paystackcontent }} onMessage={(e) => {
-                    messageReceived(e.nativeEvent?.data);
-                }} onLoadStart={() => setisLoading(true)} onLoadEnd={() => setisLoading(false)} onNavigationStateChange={onNavigationStateChange} ref={webView} cacheEnabled={false} cacheMode={'LOAD_NO_CACHE'} />
-
-                {isLoading && (<react_native_1.View>
-                    <react_native_1.ActivityIndicator size="large" color={activityIndicatorColor} />
-                </react_native_1.View>)}
-            </react_native_1.SafeAreaView>
-        </react_native_1.Modal>) : (
-             <react_native_1.View style={[{ flex: 1 }, style]}>
-             <react_native_webview_1.WebView style={[ { flex: 1 } ]} source={{ html: Paystackcontent }} onMessage={(e) => {
-                 messageReceived(e.nativeEvent?.data);
-             }} onLoadStart={() => setisLoading(true)} onLoadEnd={() => setisLoading(false)} onNavigationStateChange={onNavigationStateChange} ref={webView} cacheEnabled={false} cacheMode={'LOAD_NO_CACHE'} />
-
-             {isLoading && (<react_native_1.View>
-                 <react_native_1.ActivityIndicator size="large" color={activityIndicatorColor} />
-             </react_native_1.View>)}
-         </react_native_1.View>
-        )
-    
+var __importDefault = (this && this.__importDefault) || function (mod) {
+    return (mod && mod.__esModule) ? mod : { "default": mod };
 };
-exports.default = react_1.forwardRef(Paystack);
-//# sourceMappingURL=paystack.js.map
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.paystackProps = exports.Paystack = void 0;
+const react_native_1 = require("react-native");
+const paystack_1 = __importDefault(require("./paystack"));
+exports.Paystack = paystack_1.default;
+const paystackProps = __importStar(require("./types"));
+exports.paystackProps = paystackProps;
+exports.default = react_native_1.NativeModules.ReactNativePaystackWebviewModule;
+//# sourceMappingURL=index.js.map
