+++ lib/paystack.js
@@ -25,7 +25,7 @@ const react_native_1 = require("react-native");
 const react_native_webview_1 = require("react-native-webview");
 const helper_1 = require("./helper");
 const CLOSE_URL = 'https://standard.paystack.co/close';
-const Paystack = ({ paystackKey, billingEmail, phone, lastName, firstName, amount = '0.00', currency = 'NGN', channels = ['card'], refNumber, billingName, subaccount, handleWebViewMessage, onCancel, autoStart = false, onSuccess, activityIndicatorColor = 'green', }, ref) => {
+const Paystack = ({modal = true, style={}, paystackKey, billingEmail, phone, lastName, firstName, amount = '0.00', currency = 'NGN', channels = ['card'], refNumber, billingName, subaccount, handleWebViewMessage, onCancel, autoStart = false, onSuccess, activityIndicatorColor = 'green', }, ref) => {
     const [isLoading, setisLoading] = react_1.useState(true);
     const [showModal, setshowModal] = react_1.useState(false);
     const webView = react_1.useRef(null);
@@ -129,17 +129,29 @@ const Paystack = ({ paystackKey, billingEmail, phone, lastName, firstName, amoun
             setshowModal(false);
         }
     };
-    return (<react_native_1.Modal style={{ flex: 1 }} visible={showModal} animationType="slide" transparent={false}>
-      <react_native_1.SafeAreaView style={{ flex: 1 }}>
-        <react_native_webview_1.WebView style={[{ flex: 1 }]} source={{ html: Paystackcontent }} onMessage={(e) => {
-            messageReceived(e.nativeEvent?.data);
-        }} onLoadStart={() => setisLoading(true)} onLoadEnd={() => setisLoading(false)} onNavigationStateChange={onNavigationStateChange} ref={webView} cacheEnabled={false} cacheMode={'LOAD_NO_CACHE'}/>
+    return modal === true ?
+        (<react_native_1.Modal style={[ { flex: 1 }, style ]} visible={showModal} animationType="slide" transparent={false}>
+            <react_native_1.SafeAreaView style={{ flex: 1 }}>
+                <react_native_webview_1.WebView style={[ { flex: 1 } ]} source={{ html: Paystackcontent }} onMessage={(e) => {
+                    messageReceived(e.nativeEvent?.data);
+                }} onLoadStart={() => setisLoading(true)} onLoadEnd={() => setisLoading(false)} onNavigationStateChange={onNavigationStateChange} ref={webView} cacheEnabled={false} cacheMode={'LOAD_NO_CACHE'} />
 
-        {isLoading && (<react_native_1.View>
-            <react_native_1.ActivityIndicator size="large" color={activityIndicatorColor}/>
-          </react_native_1.View>)}
-      </react_native_1.SafeAreaView>
-    </react_native_1.Modal>);
+                {isLoading && (<react_native_1.View>
+                    <react_native_1.ActivityIndicator size="large" color={activityIndicatorColor} />
+                </react_native_1.View>)}
+            </react_native_1.SafeAreaView>
+        </react_native_1.Modal>) : (
+             <react_native_1.View style={[{ flex: 1 }, style]}>
+             <react_native_webview_1.WebView style={[ { flex: 1 } ]} source={{ html: Paystackcontent }} onMessage={(e) => {
+                 messageReceived(e.nativeEvent?.data);
+             }} onLoadStart={() => setisLoading(true)} onLoadEnd={() => setisLoading(false)} onNavigationStateChange={onNavigationStateChange} ref={webView} cacheEnabled={false} cacheMode={'LOAD_NO_CACHE'} />
+
+             {isLoading && (<react_native_1.View>
+                 <react_native_1.ActivityIndicator size="large" color={activityIndicatorColor} />
+             </react_native_1.View>)}
+         </react_native_1.View>
+        )
+    
 };
 exports.default = react_1.forwardRef(Paystack);
 //# sourceMappingURL=paystack.js.map
